+++ examples/flexbox_gap.rs
@@ -3,7 +3,7 @@ use taffy::prelude::*;
 // Creates three 20px x 20px children, evenly spaced 10px apart from each other
 // Thus the container is 80px x 20px.
 
-fn main() -> Result<(), taffy::error::TaffyError> {
+fn main() -> Result<(), taffy::TaffyError> {
     let mut taffy = Taffy::new();
 
     let child_style = Style { size: Size { width: points(20.0), height: points(20.0) }, ..Default::default() };
@@ -18,7 +18,7 @@ fn main() -> Result<(), taffy::error::TaffyError> {
 
     // Compute layout and print result
     taffy.compute_layout(root, Size::MAX_CONTENT)?;
-    taffy::debug::print_tree(&taffy, root);
+    taffy::util::print_tree(&taffy, root);
 
     Ok(())
 }