+++ src/compute/grid/alignment.rs
@@ -1,14 +1,14 @@
 //! Alignment of tracks and final positioning of items
 use super::types::GridTrack;
-use crate::axis::InBothAbsAxis;
 use crate::compute::common::alignment::compute_alignment_offset;
+use crate::geometry::InBothAbsAxis;
 use crate::geometry::{Line, Point, Rect, Size};
-use crate::layout::{Layout, SizingMode};
-use crate::math::MaybeMath;
-use crate::resolve::{MaybeResolve, ResolveOrZero};
 use crate::style::{AlignContent, AlignItems, AlignSelf, AvailableSpace, Position};
-use crate::sys::{f32_max, f32_min};
+use crate::tree::{Layout, SizingMode};
 use crate::tree::{LayoutTree, NodeId};
+use crate::util::sys::{f32_max, f32_min};
+use crate::util::MaybeMath;
+use crate::util::{MaybeResolve, ResolveOrZero};
 
 /// Align the grid tracks within the grid according to the align-content (rows) or
 /// justify-content (columns) property. This only does anything if the size of the