+++ src/compute/grid/types/cell_occupancy.rs
@@ -1,9 +1,9 @@
 //! Contains CellOccupancyMatrix used to track occupied cells during grid placement
 use super::TrackCounts;
-use crate::axis::AbsoluteAxis;
 use crate::compute::grid::OriginZeroLine;
+use crate::geometry::AbsoluteAxis;
 use crate::geometry::Line;
-use crate::sys::Vec;
+use crate::util::sys::Vec;
 use core::cmp::{max, min};
 use core::fmt::Debug;
 use core::ops::Range;