+++ src/compute/grid/types/grid_item.rs
@@ -1,17 +1,17 @@
 //! Contains GridItem used to represent a single grid item during layout
 use super::GridTrack;
-use crate::axis::AbstractAxis;
 use crate::compute::grid::OriginZeroLine;
+use crate::geometry::AbstractAxis;
 use crate::geometry::{Line, Point, Rect, Size};
-use crate::layout::SizingMode;
-use crate::math::MaybeMath;
 use crate::prelude::LayoutTree;
-use crate::resolve::{MaybeResolve, ResolveOrZero};
 use crate::style::{
     AlignItems, AlignSelf, AvailableSpace, Dimension, LengthPercentageAuto, MaxTrackSizingFunction,
     MinTrackSizingFunction, Overflow, Style,
 };
 use crate::tree::NodeId;
+use crate::tree::SizingMode;
+use crate::util::MaybeMath;
+use crate::util::{MaybeResolve, ResolveOrZero};
 use core::ops::Range;
 
 /// Represents a single grid item