+++ src/compute/mod.rs
@@ -9,25 +9,22 @@ pub(crate) mod flexbox;
 #[cfg(feature = "grid")]
 pub(crate) mod grid;
 
-use crate::error::TaffyError;
 use crate::geometry::{Point, Size};
-use crate::layout::{Layout, RunMode, SizeAndBaselines, SizingMode};
-use crate::node::Taffy;
 use crate::style::{AvailableSpace, Display};
-use crate::sys::round;
-use crate::tree::{LayoutTree, NodeId};
+use crate::tree::{Layout, LayoutTree, NodeId, RunMode, SizeAndBaselines, SizingMode, Taffy, TaffyError};
+use crate::util::sys::round;
 
 #[cfg(feature = "flexbox")]
-use self::flexbox::FlexboxAlgorithm;
+pub use self::flexbox::FlexboxAlgorithm;
 
 #[cfg(feature = "grid")]
-use self::grid::CssGridAlgorithm;
+pub use self::grid::CssGridAlgorithm;
 
 #[cfg(any(feature = "debug", feature = "profile"))]
-use crate::debug::NODE_LOGGER;
+use crate::util::debug::NODE_LOGGER;
 
 /// Updates the stored layout of the provided `node` and its children
-pub fn compute_layout(
+pub(crate) fn compute_layout(
     taffy: &mut Taffy,
     root: NodeId,
     available_space: Size<AvailableSpace>,