+++ src/lib.rs
@@ -18,33 +18,17 @@ extern crate alloc;
 #[cfg(feature = "serde")]
 extern crate serde;
 
-pub mod axis;
-pub mod cache;
-#[cfg(feature = "std")]
-#[doc(hidden)]
-pub mod debug;
-pub mod error;
+pub mod compute;
 pub mod geometry;
-pub mod layout;
-pub mod math;
-pub mod node;
 pub mod prelude;
 pub mod style;
 pub mod style_helpers;
 pub mod tree;
+pub mod util;
 
-#[cfg(feature = "random")]
-pub mod randomizable;
-
-mod compute;
-mod data;
-mod resolve;
-mod sys;
-
-pub use crate::compute::compute_layout;
 #[cfg(feature = "flexbox")]
 pub use crate::compute::flexbox::FlexboxAlgorithm;
 #[cfg(feature = "grid")]
 pub use crate::compute::grid::CssGridAlgorithm;
 pub use crate::compute::LayoutAlgorithm;
-pub use crate::node::Taffy;
+pub use crate::tree::{LayoutTree, Taffy, TaffyError, TaffyResult};